+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added - 2019-02-16
 
 - Adding `strict` mode to `object` validator.
+- Adding `throw` middleware.
 
 ## [1.7.0] - 2019-02-05